final (i hope) work for SUPPORT-178
issue is the unconditional overwriting of nodetype+model with the
default in sys::init and sys::copymodelcfginfo, which happens when
update runs on a node that's marked down or has gone down right now or
is otherwise broken when update is run.

i removed the non-working special casing for down nodes from sys::init
as it didn't cover all this-is-down cases properly (and nodeStatus() isn't
available at that time),

extended copymodelcfginfo, with type=all now overwriting type+model
only if altogether missing; then added type=overwrite which unconditionally
overwrites, and made doUpdate() use the latter once it's successfully loaded
a node's new model info on update (ie. only if node is reachable and
cooperating).

tested and works great: a new node that starts in state up gets the
correct type and model, a new node that starts dead/down gets
type=generic, update of a node while it's down (marked or defacto)
does not change the type/model, update of a node that is up overwrites
the type/model with what we've just determined from the model
selection.