workaround/fix for SUPPORT-394, gaps on one telnet host and only for ifoututil, NOT hc or inutil. reason is: rrd file had too low max set.
fixed rrd_tune_interfaces.pl: logic for fixing interface rrds was set to update if max IS EQUAL to speed-in-bytes or blank,
and not update if max is anything else.

added version, added facility to run on given nodes only, updated logging a little.
backwards/bug-compatible: sets max to TWICE the speed-in-bytes, no idea why that was done.