NMIS-80, patch updated values into config.nmis, and some more for OMK-957, flatter rrd structure
added patch_config.pl from mojo (adjusted for running with func instead of
omk::common), updated installer to provide more useful output and warnings,
to patch the non_stateful_events in as required,
and to run migrate_rrd_locations.pl if necessary.

updated nmis.pl to NOT create the unwanted hardcoded old hierarchy of
.../database/health,misc,this,that etc. rrd creation function takes care
of creating missing directories on the go.

updated migrate_rrd_locations.pl to correctly set user and group of any
new dirs, and to disregard umask for the final perms of these dirs.