OMK-1392: replace Data::UUID with UUID::Tiny
HOWEVER, this is broken as designed: if you set uuid_namespace_name,
then EVERY node will get THE SAME UUID which clearly doesn't serve any
useful purpose. that's precisely how namespaced uuids are made: the name
is the only thing controlling the uuid. solution: add the nodename to
the name (used as prefix)