SUPPORT-2260, amended runserver() to fall back to old storage data for nodeinfo if loadinfo() fails
reason is that the device name is needed/used when creating alerts/events. if loadinfo (ie. snmp) fails, then
this info is lost and nmis may create insta-orphaned incomplete events.

changed so that  if loadinfo for storage goes south, an error msg about continuning with old data is logged
and the old data is kept - for nodeinfo only, no rrd updates are performed (that would be silly).

the underlying issue is deeper; most of these data interdependencies are not safeguarded, nor do any of these
semi-high-level functions provide any error handling whatsoever...