final fix for 2b988aae65249b82358d9ac223aaa04970365939, memoryPool graphtype collision
common-cisco-memory associated one graphtype with two different sections
which cannot work, and which the new model loading code refuses to
operate on.

nmis8 code didn't hit that issue as badly as it was not relying
on model-sourced graphtype info but on nodeinfo-level graphtype data
which was the result of the higher-level logic selecting at most one
conflicting section.

resolved by cloning graph-ciscomemorypool as graph-cempmempool and
by referencing that in the cempmempool section.