nmis9: very dirty hacks in node.pl and sys.pm to make non-node graphs displayable
note that there is no way out of such a drill in, no other graph types known,
no other nodes, zero. this is inherent in the horrible non-node mode of sys,
which isn't nmis9 compatible.

also adjusted gettypeinstances to return a modeldata object under
all circumstances if want_modeldata is on - not just if things succeed.