working on OMK-4353, nmis9 periodic problem with snmp-based service checks on thor and only on thor
cleaned up logic somewhat; all indications were that some dud data in catchall for that node caused a broken logic branch
to be entered (nothing in the logs, no event but status 0in rrd and timed_data service entry
while LOTS of correct processes were shown in timed_data snmp_services)

branch was 'snmp was down before service check' which caused INCORRECT information to be written to rrd and db,
but no service down event raising (so not a 'real' proper 'service down' situation); logic was legacy/semi-useless
meant to adjust the view file in that case but did much much more.

that branch is now gone, and i'm hoping for correct operation on thor now.