OMK-4609, nmis9: getintfdata() ran into assert too easily
issue triggered by tun interfaces being shutdown and recreated (ie. restart openvpn)
where the kernel leaves the goner interface alive for a while (with admin=operstatus down).

if the interface is then actually removed while getintfdata() still works, then sys::loadinfo() is asked
to retrieve a goner's interface info and that function DOES NOT report nosuchinstance as a problem.

this caused a) clashing phantom interfaces for multiple different ifindices  but all
called noSuchInstance (which causes other internal safety-net assertions)

and b) an already updated broken/goner interface would get mis-identified as being in need of a
stage-7 update because its ifstatus stinks of a transition.

all of that is fixed now; getIntfInfo() was taught to not return nosuchinstance interfaces,
getIntfData() was taught to deal with interfaces being removed during its run, and to not consider
already stage-4-updated interfaces for stage 7 under any circumstances.