nmis9, OMK-4819, teach nmisng::db::find to trigger and catch cursor-caused late exceptions
basically reverts f2b884 and replaces it with a smarter solution, ie.
find() now uses has_next() to tickle the cursor before returning it; if that works, fine.
but if it throws an exception, set error_string and return undef.

the next_result and all_results wrappers are not necessary and have been removed.