nmisng: make events system use db instead of file, may have issues
Create an event object with the known data, if you think it's new call save
and it will be added/updated. Call load or exists to find out if it's in the db
Functions that work on multiple events are in Events.pm, it also holds an API
that mimics Comapt::NMIS.

Note: event->{current} is now event->{active}, and current/historic folders is
now event->{historic}

ack and all other true/falses have been changed to 0/1, getbool still works
on the new values but is not required.

New tests for Event object in t_event.pl, does not have 100% but does try and
cover off most things. t_notify.pl also test the event system but is not as
thorough, just checks to make sure the right number of events is created and
 that they properly go from active->1/0 and then historic