nmis9, OMK-4982, nmis should treat misconfigured services as down, not skip them
ie. program/script files that don't exist or cannot be executed, port-type services
without valid port and service and so on.

in such a case, a temporary service status text with the complaint is
created (regardless of whether the service in question normally can/should collect
'command outputs') which is shown in the gui as the  'last status text'.

manual port of 9760cc to nmis9