nmis9, omk-5888, patch_config race condition clobbers config with -b
first problem, rename() of config file just before it's about to
be written cannot work as writehashtofile calls loadconftable() which
checks the config file for existence and falls back to the one from
conf-default if n/a.

another catch-22 exists for almost all use of nmisng::util::dbg() which
requires a full nmisng object, which in turn requires a db connection,
which is not in any way guaranteeable when ancillary code is run (e.g.
when patch_config uses writehashtofile).

dbg() has to be removed for a general release and code in need
of debug-logging must learn to use the nmisng::log module.

for now i've removed the unnecessary call to dbg() from writehashtofile.