nmis9, OMK-4358, work around dots in node override keys
mongodb < 3.6 doesn't support such keys. for now override keys are
base64 encoded and prefixed with ==, and decoded on the fly.

only users of the raw data of get_nodes_model (ie. not using ->objects)
need to know about this. currently that's node_admin,
compat::nmis::loadlocalnodetable and  compat::nmis::loadnodetable, all of which
have been updated.