nmis9, OMK-5961, get rid of var/nmis-fping.json and update db instead
nmisd's fping worker now saves ping values as timed data for concept 'ping'.
the nmis-fping state file is no longer generated or referenced.

cgi-bin/network.pl was updated to get the last_ping info from the db, and no
longer looks at the catchall for ping info

nmisng::node::pingable was updated to get the ping info from the timed data
instead of the nmis-fping state file, and to not park ping-related info in the
catchall.