nmis9, towards OMK-5975, smart/minimal save for inventory objects
feature is complete but INSTUFFICIENTLY TESTED at this time!

the inventory code was updated to keep track of what changes to the object
dirty up the state; where possible save will then only update the things that
have changed between the object having been loaded and the time of save.

this should avoid a number of race conditions, especially with data_live use
(which is discouraged); additionally it'll reduce the number of
'nuisance updates' in the database (ie. where only expire_at and lastupdate
have changed).