nmis9, SUPPORT-6291, bug in collection scheduling logic causes lots of load
issue was triggered by reusing catchall inventories in data_live mode
across multiple save operations. mistake was in save which reset the
baseline original data, reasoning: stuff's now on disk and hence safe so we
can go back to clean/initial state...which is correct in non-live mode.
in data-live mode as used with sys & co. there are multiple holders
of the same data_live ref that simply continue to update the data.

the buggy version would then establish a new baseline of original data
on the next call to data_live, which however caused it to lose any changes
that happened inbetween the save() and this data_live() call.