No SNMP down when mib name can't be found, add placeholder for systemHealth/sys
If a systemHealth indexed value cannot be converted into an oid do not trigger
snmp down, log an error. If there is no value log the error and continue.

If placeholder is set for the systemHealth sys section do not expect and rrd
section, plugin will be expected to fill it in.