OMK-10213 - parsed strings quoted with single quotes
Some replacement strings had quotes in them which was causing issues.
Switching to single quotes so nothing in the replacement string can
be evaluated. Removing any single quotes before doing that so the
single quotes are not interrupted.

The changes to the strings are not saved, this is only done just
before evaluation. I do not see any reason we should not just be
quoting all the time instead of running a regex on it, discussion
on this is needed.