Always treat Inventory index data attribute as a string
In cases where an index_regex is used we can end up in a place
where perl interpolation (and a trip to/from the db) ends up
changing 1.0 to 1 which breaks the snmp request. forcing
data->index to a string solves the issue and should be safe
as no number operations are done on it.

The value in the path will still be reduced to 1, this should not
be an issue as both the save and query use the same code to
numify beforehand. There is also no risk of an inventory concept
needing 1.0 and 1 as indexes.