Remove ifDescrInfo and replace with interface_by_ifDescr
ifDescrInfo loads all interfaces for a node and arrages them
by ifDescr, this is extremely inefficient on larger installs or
nodes with many interfaces.

This is replaced by a search for an interface with a specific
ifDescr for a node.

This is a hangover from when nmis8 used files and all interfaces
would have been loaded anyway.