Track timed collections and only create indexes when required
The goal is to significanly reduce the number of times create index
is called for timed collections.

Track the timed_* collections that currently exist, only call create
index when a collection is requested that the object doesn't know about.

This functionality is really only needed for nmisd which is why it has been
added the way it was, this does not need to run every time a new nsming
object is made