Turn off updating historic interfaces in collect
Some devices end up with many historic interfaces, attempting to
update them all every collect cycle causes many issues.

This code was added to deal with a node that was reset between
updates, an interface disappeared (was marked historic) and came
back (but was still historic until the next update).

We may want a separate job to check historic inventory are still
non-existant instead.  Or trigger an update when a node is found
to be reset.

Another option may be to use ifTableLastChange to decide about
more things