wqBetter handling for interfaces with duplicate ifDescr
We are seeing this in many VPN links, we rely on ifDescr to be invariant
to handle ifIndex changes. This finds all dups, makes sure they are all
renamed and checks for that renaming when collecting.

There are still some issues with old inventory sticking around and not
being marked historic and fading away.