Oids in models should not start with a period, find and remove them
This is done before the model is cached so that it does not have to
happen often.

Some agents (or possibly our SNMP lib) returns all values with
preceeding "." if one of the oid's supplied has a preceeding "."
This breaks all other values in that request because the mapping
of request oid -> return oid does not work because the return oid
keys have changed. To prevend this we just don't allow them to start
with a "."