NMIS DB optimisation - send inventory as arg for parseString
Small change, some functions that end up calling parse string
already have their inventory loaded, so send it along as an
argument so parseString does not load it again. If it's not
there parseString will load it anyway, and it only uses ->data
which takes a copy of the data that it currently has.

The risk here is that parseString is now using newer data than
it was before (instead of loading what is in the db it's using
whatever is commited to the current inventory). This is likely
 not actually new as most functions set the data at the end.
parseString generally uses "update" data and not collect data
so it's most likely it's the same info anyway.