Very crude first go at logging
if it's turned on and log file exists it will write to it, this
should be expanded to capture more/better data. it should also buffer
this up and write in chunks.

would be good to have a way to enable/disable for a collect or update
so we can track db request changes over time for a fixed type of node
or nodes