No SNMP connection without community or username, reachability stats
SNMP down is always made even if no snmp is wanted, like when wmi
is enabled. Change requires a community string or snmpv3 username
in order for snmp to be attempted.

reachability was only looking at snmpresult, now it looks at both,
taking the worse of the two if they are defined (so wmi only will
get 100 reachability if ping and wmi are good)