Make sure up/down events are created, if no creds for anything snmp enables
Sys now only tries polling engines that have credetials/configuration. If
no configuration for anything is found SNMP is enabled so one of the engines
gives a down error (this is done to mimic previous behaviour). Otherwise
we end up with a node that is collect:true but no event for any engine
down, if this is intentional collect should be set to false so a pingonly
model is used.