No SNMP connection without community or username, reachability stats (#85)
* No SNMP connection without community or username, reachability stats

SNMP down is always made even if no snmp is wanted, like when wmi
is enabled. Change requires a community string or snmpv3 username
in order for snmp to be attempted.

reachability was only looking at snmpresult, now it looks at both,
taking the worse of the two if they are defined (so wmi only will
get 100 reachability if ping and wmi are good)

* Make sure up/down events are created, if no creds for anything snmp enables

Sys now only tries polling engines that have credetials/configuration. If
no configuration for anything is found SNMP is enabled so one of the engines
gives a down error (this is done to mimic previous behaviour). Otherwise
we end up with a node that is collect:true but no event for any engine
down, if this is intentional collect should be set to false so a pingonly
model is used.

---------

Co-authored-by: Mark Dueck <mark.dueck@firstwave.com>