update wmic binary for json output bug, add timeout for wmi queries
wmic binary was generating invalid json for requests with a large
amount of output.

added timeout for devices that allow a connection but do not respond
properly. it was defaulting to 300s, now gets cut off at 20s. the
timeout to the wmic binary is set but the alarm goes off first