Clear active jobs on startup, fix saving log output to queue and log
OMK-11180 - nmisd now clears job queue of all active jobs on startup.

OMK-11791 - new function to capture and log data so nmis.log gets the output
and the job queue has the data saved to it as well. part of this was added
before, this fixes it up, seems to be working properly now