Add nmis8 nodesum support and new event for invalid model file
nodesum support required grabbing a bit of extra data from the collect cycle and putting it into the catchall

Event "Model File Invalid" will now be associated with a node if NMIS determines a specific Model file should be loaded for a node and it failes to load. This event will only be associated on update and cleared on update.